fix: ログインページにサインアップフォームを統合（本家HN準拠）
